requirements.txt
discord.py
pinata_python
dotenv
web3